requirements.txt
Django==4.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
Pillow==10.2.0
python-decouple==3.8
psycopg2==2.9.9
django-filter==25.1
setuptools==80.9.0


